refactor(rust): prop strings mod
