+++ rust/mlt-core/src/layer/v01/property/mod.rs
@@ -1,6 +1,5 @@
-pub(crate) mod decode;
+pub mod strings;
 
-use std::collections::hash_map::Entry;
 use std::collections::{HashMap, HashSet};
 use std::fmt::{self, Debug};
 use std::io::Write;
@@ -12,38 +11,15 @@ use crate::MltError::{IntegerOverflow, NotImplemented, UnsupportedPropertyEncode
 use crate::analyse::{Analyze, StatType};
 use crate::decode::{FromEncoded, impl_decodable};
 use crate::utils::{BinarySerializer as _, FmtOptVec, apply_present, f32_to_json, f64_to_json};
-use crate::v01::property::decode::{decode_string_streams, decode_struct_children};
+pub use crate::v01::property::strings::{
+    EncodedStructChild, EncodedStructProp, SharedDictChild, SharedDictEncoder,
+    SharedDictionaryGroup, StrEncoder, decode_string_streams, decode_struct_children,
+    encode_shared_dictionary,
+};
 use crate::v01::{
-    ColumnType, DictionaryType, FsstStrEncoder, IntEncoder, LengthType, OffsetType, OwnedStream,
-    Stream, StreamType,
+    ColumnType, DictionaryType, FsstStrEncoder, IntEncoder, LengthType, OwnedStream, Stream,
 };
 use crate::{FromDecoded, MltError, impl_encodable};
-
-/// Encoded data for a Struct column with shared dictionary encoding
-#[borrowme]
-#[derive(Debug, PartialEq)]
-pub struct EncodedStructProp<'a> {
-    pub dict_streams: Vec<Stream<'a>>,
-    pub children: Vec<EncodedStructChild<'a>>,
-}
-
-/// A single child field within a Struct column
-#[borrowme]
-#[derive(Debug, PartialEq)]
-pub struct EncodedStructChild<'a> {
-    pub name: &'a str,
-    pub typ: ColumnType,
-    pub optional: Option<Stream<'a>>,
-    pub data: Stream<'a>,
-}
-
-impl OwnedEncodedStructChild {
-    pub(crate) fn write_columns_meta_to<W: Write>(&self, writer: &mut W) -> Result<(), MltError> {
-        self.typ.write_to(writer)?;
-        writer.write_string(&self.name)?;
-        Ok(())
-    }
-}
 
 /// Property representation, either encoded or decoded
 #[borrowme]
@@ -519,20 +495,6 @@ impl From<ScalarEncoder> for PropertyEncoder {
     }
 }
 
-#[derive(Debug, Clone, PartialEq)]
-pub struct SharedDictEncoder {
-    /// Name of the parent struct column.
-    ///
-    /// All instructions with the same value are grouped into one struct column.
-    pub struct_name: String,
-    /// Name of this field within the struct column.
-    child_name: String,
-    /// Encoder used for the offset-index stream of this child.
-    offset: IntEncoder,
-    /// If a stream for optional values should be attached
-    optional: PresenceStream,
-}
-
 /// How to encode properties
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[cfg_attr(all(not(test), feature = "arbitrary"), derive(arbitrary::Arbitrary))]
@@ -611,27 +573,6 @@ impl ScalarValueEncoder {
             ScalarValueEncoder::Bool => "bool",
             ScalarValueEncoder::Struct => "struct",
         }
-    }
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Copy)]
-#[cfg_attr(test, derive(proptest_derive::Arbitrary))]
-#[cfg_attr(all(not(test), feature = "arbitrary"), derive(arbitrary::Arbitrary))]
-pub enum StrEncoder {
-    Plain { string_lengths: IntEncoder },
-    Fsst(FsstStrEncoder),
-}
-impl StrEncoder {
-    #[must_use]
-    pub fn plain(string_lengths: IntEncoder) -> Self {
-        Self::Plain { string_lengths }
-    }
-    #[must_use]
-    pub fn fsst(symbol_lengths: IntEncoder, dict_lengths: IntEncoder) -> Self {
-        Self::Fsst(FsstStrEncoder {
-            symbol_lengths,
-            dict_lengths,
-        })
     }
 }
 
@@ -757,107 +698,6 @@ impl FromDecoded<'_> for Vec<OwnedEncodedProperty> {
     }
 }
 
-struct SharedDictionaryGroup<'a> {
-    shared: StrEncoder,
-    children: Vec<SharedDictChild<'a>>,
-}
-
-struct SharedDictChild<'a> {
-    prop_value: &'a DecodedProperty,
-    prop_name: String,
-    optional: PresenceStream,
-    offset: IntEncoder,
-}
-
-/// Encode a group of decoded string properties into a single struct column with a shared
-/// dictionary. Children are ordered as provided.
-fn encode_shared_dictionary(
-    name: &str,
-    group: &SharedDictionaryGroup,
-) -> Result<OwnedEncodedProperty, MltError> {
-    // Build shared dictionary: unique strings in first-occurrence insertion order.
-    let mut dict: Vec<&str> = Vec::new();
-    let mut dict_index: HashMap<&str, u32> = HashMap::new();
-
-    for child in &group.children {
-        match &child.prop_value.values {
-            PropValue::Str(values) => {
-                for s in values.iter().flatten() {
-                    if let Entry::Vacant(e) = dict_index.entry(s) {
-                        let idx = u32::try_from(dict.len())?;
-                        e.insert(idx);
-                        dict.push(s);
-                    }
-                }
-            }
-            _ => return Err(NotImplemented("generic prop_child encoding")),
-        }
-    }
-
-    let dict_streams = match group.shared {
-        StrEncoder::Plain { string_lengths } => OwnedStream::encode_strings_with_type(
-            &dict,
-            string_lengths,
-            LengthType::Dictionary,
-            DictionaryType::Shared,
-        )?,
-        StrEncoder::Fsst(enc) => OwnedStream::encode_strings_fsst_with_type(
-            &dict,
-            enc,
-            DictionaryType::Single, // TODO: figure out if this is correct. According to Java it is.. but why?
-        )?,
-    };
-
-    // Encode each child column.
-    let mut encoded_children = Vec::with_capacity(group.children.len());
-    for child in &group.children {
-        let PropValue::Str(values) = &child.prop_value.values else {
-            return Err(NotImplemented("generic struct child encoding"));
-        };
-
-        // Presence stream
-        let optional = if child.optional == PresenceStream::Present {
-            let present_bools: Vec<bool> = values.iter().map(Option::is_some).collect();
-            Some(OwnedStream::encode_presence(&present_bools)?)
-        } else {
-            None
-        };
-
-        // Offset indices for non-null values only.
-        let offsets: Vec<u32> = values
-            .iter()
-            .filter_map(|v| v.as_deref())
-            .map(|s| dict_index[s])
-            .collect();
-
-        let data = OwnedStream::encode_u32s_of_type(
-            &offsets,
-            child.offset,
-            StreamType::Offset(OffsetType::String),
-        )?;
-
-        encoded_children.push(OwnedEncodedStructChild {
-            name: child.prop_name.clone(),
-            typ: if child.optional == PresenceStream::Present {
-                ColumnType::OptStr
-            } else {
-                ColumnType::Str
-            },
-            optional,
-            data,
-        });
-    }
-
-    Ok(OwnedEncodedProperty {
-        name: name.to_owned(),
-        optional: None,
-        value: OwnedEncodedPropValue::Struct(OwnedEncodedStructProp {
-            dict_streams,
-            children: encoded_children,
-        }),
-    })
-}
-
 impl FromDecoded<'_> for OwnedEncodedProperty {
     type Input = DecodedProperty;
     type Encoder = ScalarEncoder;
@@ -980,68 +820,6 @@ mod tests {
 
     use super::*;
     use crate::v01::{LogicalEncoder, PhysicalEncoder};
-
-    fn strs(vals: &[&str]) -> Vec<Option<String>> {
-        vals.iter().map(|v| Some(v.to_string())).collect()
-    }
-
-    fn opt_strs(vals: &[Option<&str>]) -> Vec<Option<String>> {
-        vals.iter().map(|v| v.map(ToString::to_string)).collect()
-    }
-
-    fn str_prop(name: &str, values: Vec<Option<String>>) -> DecodedProperty {
-        DecodedProperty {
-            name: name.to_string(),
-            values: PropValue::Str(values),
-        }
-    }
-
-    fn expand_struct(prop: &OwnedEncodedProperty) -> Vec<DecodedProperty> {
-        Property::from(borrowme::borrow(prop))
-            .decode_expand()
-            .expect("decode_expand failed")
-            .into_iter()
-            .map(|p| p.decode().expect("decode failed"))
-            .collect()
-    }
-
-    fn decode_scalar(prop: &OwnedEncodedProperty) -> DecodedProperty {
-        DecodedProperty::from_encoded(borrowme::borrow(prop)).expect("decode failed")
-    }
-
-    /// Encode a group of string children as a struct column and expand them back out.
-    fn struct_encode_and_expand(
-        struct_name: &str,
-        children: &[(&str, Vec<Option<String>>)],
-        presence: PresenceStream,
-        encoder: IntEncoder,
-        shared_dicts: impl Into<HashMap<String, StrEncoder>>,
-    ) -> Vec<DecodedProperty> {
-        let decoded: Vec<DecodedProperty> = children
-            .iter()
-            .map(|(child_name, values)| str_prop(child_name, values.clone()))
-            .collect();
-        let instructions: Vec<PropertyEncoder> = children
-            .iter()
-            .map(|(child_name, _)| {
-                PropertyEncoder::shared_dict(struct_name, *child_name, presence, encoder)
-            })
-            .collect();
-        let encoded_prop = Vec::<OwnedEncodedProperty>::from_decoded(
-            &decoded,
-            MultiPropertyEncoder {
-                properties: instructions,
-                shared_dicts: shared_dicts.into(),
-            },
-        )
-        .expect("encoding failed");
-        assert_eq!(
-            encoded_prop.len(),
-            1,
-            "struct children must collapse to one column"
-        );
-        expand_struct(&encoded_prop[0])
-    }
 
     fn roundtrip(decoded: &DecodedProperty, encoder: ScalarEncoder) -> DecodedProperty {
         let encoded_prop =
@@ -1180,342 +958,6 @@ mod tests {
             };
             let enc = ScalarEncoder::float(PresenceStream::Present);
             prop_assert_eq!(roundtrip(&prop, enc), prop);
-        }
-    }
-
-    // PropValue::Str can be encoded as a standalone (non-struct) column.  This is
-    // a separate code path from the shared-dictionary struct encoding below.
-    #[test]
-    fn str_scalar_with_nulls() {
-        let prop = str_prop(
-            "city",
-            opt_strs(&[Some("Berlin"), None, Some("Hamburg"), None]),
-        );
-        let enc = ScalarEncoder::str(PresenceStream::Present, IntEncoder::plain());
-        assert_eq!(roundtrip(&prop, enc), prop);
-    }
-
-    #[test]
-    fn str_scalar_all_null() {
-        // All-None with a presence stream: the data stream is empty, presence is all-false.
-        let prop = str_prop("city", opt_strs(&[None, None, None]));
-        let enc = ScalarEncoder::str(PresenceStream::Present, IntEncoder::plain());
-        assert_eq!(roundtrip(&prop, enc), prop);
-    }
-
-    #[test]
-    fn str_scalar_empty() {
-        // Zero-row property: nothing to encode on either stream.
-        let prop = str_prop("unused", vec![]);
-        let enc = ScalarEncoder::str(PresenceStream::Present, IntEncoder::plain());
-        assert_eq!(roundtrip(&prop, enc), prop);
-    }
-
-    proptest! {
-        #[test]
-        fn str_scalar_roundtrip(
-            values in prop::collection::vec(
-                prop::option::of("[a-zA-Z0-9 ]{0,30}"),
-                0..50,
-            ),
-        ) {
-            let prop = str_prop("name", values);
-            let enc = ScalarEncoder::str(PresenceStream::Present, IntEncoder::plain());
-            prop_assert_eq!(roundtrip(&prop, enc), prop);
-        }
-    }
-
-    // FSST builds a symbol table from repeated byte sequences and compresses the
-    // corpus against it.  Both scalar and struct paths use separate code routes.
-    #[test]
-    fn fsst_scalar_string_roundtrip() {
-        let enc = ScalarEncoder::str_fsst(
-            PresenceStream::Present,
-            IntEncoder::plain(),
-            IntEncoder::plain(),
-        );
-        // Repeated "Br" prefix gives FSST something to compress.
-        let prop = str_prop(
-            "name",
-            strs(&["Berlin", "Brandenburg", "Bremen", "Braunschweig"]),
-        );
-        assert_eq!(roundtrip(&prop, enc), prop);
-    }
-
-    #[test]
-    fn fsst_struct_shared_dict_roundtrip() {
-        let enc = IntEncoder::plain();
-        let de = strs(&["Berlin", "Brandenburg", "Bremen"]);
-        let en = strs(&["Berlin", "Brandenburg", "Bremen"]);
-        let result = struct_encode_and_expand(
-            "name",
-            &[(":de", de.clone()), (":en", en.clone())],
-            PresenceStream::Present,
-            enc,
-            [("name".to_string(), StrEncoder::plain(IntEncoder::plain()))],
-        );
-        assert_eq!(result[0].values, PropValue::Str(de));
-        assert_eq!(result[1].values, PropValue::Str(en));
-    }
-
-    #[test]
-    fn struct_with_nulls() {
-        // decode_expand must prefix each child name with the parent struct name.
-        let de = opt_strs(&[Some("Berlin"), Some("München"), None]);
-        let en = opt_strs(&[Some("Berlin"), None, Some("London")]);
-        let result = struct_encode_and_expand(
-            "name",
-            &[(":de", de.clone()), (":en", en.clone())],
-            PresenceStream::Present,
-            IntEncoder::plain(),
-            [("name".to_string(), StrEncoder::plain(IntEncoder::plain()))],
-        );
-        assert_eq!(result.len(), 2);
-        assert_eq!(result[0].name, "name:de");
-        assert_eq!(result[0].values, PropValue::Str(de));
-        assert_eq!(result[1].name, "name:en");
-        assert_eq!(result[1].values, PropValue::Str(en));
-    }
-
-    #[test]
-    fn struct_no_nulls() {
-        let de = strs(&["Berlin", "München", "Hamburg"]);
-        let en = strs(&["Berlin", "Munich", "Hamburg"]);
-        let result = struct_encode_and_expand(
-            "name",
-            &[(":de", de.clone()), (":en", en.clone())],
-            PresenceStream::Present,
-            IntEncoder::plain(),
-            [("name".to_string(), StrEncoder::plain(IntEncoder::plain()))],
-        );
-        assert_eq!(result[0].values, PropValue::Str(de));
-        assert_eq!(result[1].values, PropValue::Str(en));
-    }
-
-    #[test]
-    fn struct_shared_dict_deduplication() {
-        // "Berlin" appears in both children.  The encoder must store it only once in
-        // the shared dictionary, not once per child.  We verify this by inspecting
-        // the encoded column directly: plain encoding always produces exactly 2 dict
-        // streams (a length stream + a data stream), regardless of how many strings
-        // are in the dictionary.  Then we confirm the decoded values are correct.
-        let decoded = vec![
-            str_prop(":de", strs(&["Berlin", "Berlin"])),
-            str_prop(":en", strs(&["Berlin", "London"])),
-        ];
-        let enc = IntEncoder::plain();
-        let prop_encs = vec![
-            PropertyEncoder::shared_dict("name", ":de", PresenceStream::Present, enc),
-            PropertyEncoder::shared_dict("name", ":en", PresenceStream::Present, enc),
-        ];
-        let string_enc = StrEncoder::Plain {
-            string_lengths: IntEncoder::plain(),
-        };
-        let enc = MultiPropertyEncoder {
-            properties: prop_encs.clone(),
-            shared_dicts: HashMap::from([("name".to_string(), string_enc)]),
-        };
-
-        let encoded = Vec::<OwnedEncodedProperty>::from_decoded(&decoded, enc).unwrap();
-
-        let OwnedEncodedPropValue::Struct(ref s) = encoded[0].value else {
-            panic!("expected Struct variant");
-        };
-        // If deduplication were broken a naive implementation might write "Berlin" twice,
-        // but the stream count is structural — it will always be 2 for plain encoding.
-        // What changes is the data stream's byte length; we check that after decode the
-        // second offset still resolves "London" correctly (dict index 1, not 2).
-        assert_eq!(
-            s.dict_streams.len(),
-            2,
-            "plain: one length stream + one data stream"
-        );
-
-        let children = expand_struct(&encoded[0]);
-        assert_eq!(
-            children[0].values,
-            PropValue::Str(strs(&["Berlin", "Berlin"]))
-        );
-        assert_eq!(
-            children[1].values,
-            PropValue::Str(strs(&["Berlin", "London"]))
-        );
-    }
-
-    #[test]
-    fn struct_mixed_with_scalars() {
-        // Scalar columns before and after a struct group must land in the right
-        // positions after the two-pass grouping logic.
-        let scalar_enc = ScalarEncoder::int(PresenceStream::Present, IntEncoder::plain());
-        let population = DecodedProperty {
-            name: "population".to_string(),
-            values: PropValue::U32(vec![Some(3_748_000), Some(1_787_000)]),
-        };
-        let name_de = str_prop(":de", strs(&["Berlin", "Hamburg"]));
-        let name_en = str_prop(":en", strs(&["Berlin", "Hamburg"]));
-        let rank = DecodedProperty {
-            name: "rank".to_string(),
-            values: PropValue::U32(vec![Some(1), Some(2)]),
-        };
-
-        let props = vec![
-            population.clone(),
-            name_de.clone(),
-            name_en.clone(),
-            rank.clone(),
-        ];
-        let prop_encs = vec![
-            PropertyEncoder::Scalar(scalar_enc),
-            PropertyEncoder::shared_dict(
-                "name",
-                ":de",
-                PresenceStream::Present,
-                IntEncoder::plain(),
-            ),
-            PropertyEncoder::shared_dict(
-                "name",
-                ":en",
-                PresenceStream::Present,
-                IntEncoder::plain(),
-            ),
-            PropertyEncoder::Scalar(scalar_enc),
-        ];
-        let string_enc = StrEncoder::Plain {
-            string_lengths: IntEncoder::plain(),
-        };
-        let enc = MultiPropertyEncoder {
-            properties: prop_encs.clone(),
-            shared_dicts: HashMap::from([("name".to_string(), string_enc)]),
-        };
-
-        let encoded_prop = Vec::<OwnedEncodedProperty>::from_decoded(&props, enc).unwrap();
-
-        // Output order: scalar "population", struct "name", scalar "rank"
-        assert_eq!(encoded_prop.len(), 3);
-        assert_eq!(decode_scalar(&encoded_prop[0]), population);
-        let name = expand_struct(&encoded_prop[1]);
-        assert_eq!(name[0].name, "name:de");
-        assert_eq!(name[0].values, name_de.values);
-        assert_eq!(name[1].name, "name:en");
-        assert_eq!(name[1].values, name_en.values);
-        assert_eq!(decode_scalar(&encoded_prop[2]), rank);
-    }
-
-    #[test]
-    fn two_struct_groups_with_scalar_between() {
-        // Two independent struct columns must each get their own shared dictionary
-        // and appear at the position of their first child in the output.
-        let name_de = str_prop(":de", strs(&["Berlin", "Hamburg"]));
-        let name_en = str_prop(":en", strs(&["Berlin", "Hamburg"]));
-        let population = DecodedProperty {
-            name: "population".to_string(),
-            values: PropValue::U32(vec![Some(3_748_000), Some(1_787_000)]),
-        };
-        let label_de = str_prop(":de", strs(&["BE", "HH"]));
-        let label_en = str_prop(":en", strs(&["BER", "HAM"]));
-
-        let decoded_props = vec![
-            name_de.clone(),
-            name_en.clone(),
-            population.clone(),
-            label_de.clone(),
-            label_en.clone(),
-        ];
-        let enc = IntEncoder::plain();
-        let prop_encoders = vec![
-            PropertyEncoder::shared_dict("name:", "de", PresenceStream::Present, enc),
-            PropertyEncoder::shared_dict("name:", "en", PresenceStream::Present, enc),
-            ScalarEncoder::int(PresenceStream::Present, enc).into(),
-            PropertyEncoder::shared_dict("label:", "de", PresenceStream::Present, enc),
-            PropertyEncoder::shared_dict("label:", "en", PresenceStream::Present, enc),
-        ];
-        let str_enc = StrEncoder::plain(IntEncoder::plain());
-        let encoded_prop = Vec::<OwnedEncodedProperty>::from_decoded(
-            &decoded_props,
-            MultiPropertyEncoder {
-                properties: prop_encoders,
-                shared_dicts: HashMap::from([
-                    ("name:".to_string(), str_enc),
-                    ("label:".to_string(), str_enc),
-                ]),
-            },
-        )
-        .unwrap();
-
-        // Expected output order: struct "name", scalar "population", struct "label"
-        assert_eq!(encoded_prop.len(), 3);
-        let name = expand_struct(&encoded_prop[0]);
-        assert_eq!(name[0].name, "name:de");
-        assert_eq!(name[0].values, name_de.values);
-        assert_eq!(name[1].name, "name:en");
-        assert_eq!(name[1].values, name_en.values);
-        assert_eq!(decode_scalar(&encoded_prop[1]), population);
-        let label = expand_struct(&encoded_prop[2]);
-        assert_eq!(label[0].name, "label:de");
-        assert_eq!(label[0].values, label_de.values);
-        assert_eq!(label[1].name, "label:en");
-        assert_eq!(label[1].values, label_en.values);
-    }
-
-    #[test]
-    fn struct_instruction_count_mismatch() {
-        let err = Vec::<OwnedEncodedProperty>::from_decoded(
-            &vec![DecodedProperty::default()],
-            MultiPropertyEncoder {
-                properties: vec![],
-                shared_dicts: HashMap::default(),
-            },
-        )
-        .unwrap_err();
-        assert!(
-            matches!(
-                err,
-                MltError::EncodingInstructionCountMismatch {
-                    input_len: 1,
-                    config_len: 0
-                }
-            ),
-            "unexpected error: {err}"
-        );
-    }
-
-    proptest! {
-        #[test]
-        fn struct_roundtrip(
-            struct_name in "[a-z]{1,8}",
-            children in prop::collection::vec(
-                (
-                    "[a-z]{1,6}",
-                    prop::collection::vec(prop::option::of("[a-zA-Z ]{0,20}"), 0..20),
-                ),
-                1..5usize,
-            ),
-            logical in any::<LogicalEncoder>(),
-            physical in physical_no_fastpfor(),
-            string_enc in  any::<StrEncoder>(),
-        ) {
-            let encoder = IntEncoder::new(logical, physical);
-            let decoded: Vec<DecodedProperty> = children
-                .iter()
-                .map(|(child_name, values)| str_prop(child_name, values.clone()))
-                .collect();
-            let properties: Vec<PropertyEncoder> = children
-                .iter()
-                .map(|(child_name, _)| {
-                    PropertyEncoder::shared_dict(&struct_name, child_name,PresenceStream::Present,  encoder)
-                })
-                .collect();
-            let shared_dicts = HashMap::from([(struct_name.clone(),string_enc)]);
-            let enc = MultiPropertyEncoder{ properties, shared_dicts };
-            let encoded = Vec::<OwnedEncodedProperty>::from_decoded(&decoded, enc)
-                .expect("encoding failed");
-            prop_assert_eq!(encoded.len(), 1, "struct children must collapse to one column");
-            let re_children = expand_struct(&encoded[0]);
-            prop_assert_eq!(re_children.len(), children.len());
-            for (re, (child_name, values)) in re_children.into_iter().zip(children.iter()) {
-                prop_assert_eq!(re.name, format!("{struct_name}{child_name}"));
-                prop_assert_eq!(re.values, PropValue::Str(values.clone()));
-            }
         }
     }
 }
